+++ src/refactoring/hooks/admin/useEditingProduct.ts
@@ -1,15 +1,8 @@
 import { useState } from 'react';
 import { Discount, Product } from '../../../types';
-import { getFormattedValue, updateOpenProductIds } from '../../helpers';
+import { getFormattedValue } from '../../helpers';
 
-export const useAdmin = () => {
-  // 상품 상세 토글
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => updateOpenProductIds(prev, productId));
-  };
-
-  // 상품 정보 수정
+export const useEditingProduct = () => {
   const [editingProduct, setEditingProduct] = useState<Product | null>(null);
   const updateEditingProduct = (product: Product | null) => {
     setEditingProduct(product);
@@ -26,9 +19,6 @@ export const useAdmin = () => {
   };
 
   return {
-    openProductIds,
-    toggleProductAccordion,
-
     editingProduct,
     updateEditingProduct,
     handleEditingProductUpdate,
