+++ src/refactoring/hooks/admin/useOpenProductIds.ts
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+import { updateOpenProductIds } from '../../helpers';
+
+export const useOpenProductIds = () => {
+  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
+  const toggleProductAccordion = (productId: string) => {
+    setOpenProductIds((prev) => updateOpenProductIds(prev, productId));
+  };
+
+  return {
+    openProductIds,
+    toggleProductAccordion,
+  };
+};
