+++ src/refactoring/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from './cart/useCart.ts';
-export * from './admin/useAdmin.ts';
+export * from './admin/useEditingProduct.ts';
 export * from './admin/useNewCoupon.ts';
 export * from './admin/useNewDiscount.ts';
 export * from './admin/useNewProduct.ts';
